requirements.txt
yfinance>=0.2.36
openai>=1.12.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
